SphericalGeometry.jl

SphericalGeometry.SphericalGeometry — Module

A Julia Package for handling Spherical Geometry.

It includes the calculation of:

The distance to a point, line, arc, multi-arc, or polygon border

The intersection points between lines, arcs, multi-arcs, and polygon borders.

The self intersection points of multi-arcs and polygon borders

And testing if:

a point, arc or multi-arc is inside a polygon

a point is on a line, arc, multi-arc or polygon border

Types

SphericalGeometry.Point — Type

Point type with latitude ϕ [deg] and longitude λ [deg]

Missing docstring.

Missing docstring for MultiPoint. Check Documenter's build log for details.

SphericalGeometry.Arc — Type

Arc type with start position pos₁ [deg] and end position pos₂ [deg]

Missing docstring.

Missing docstring for MultiArc. Check Documenter's build log for details.

SphericalGeometry.Line — Type

Line type with point [deg] and azimuth [deg]

SphericalGeometry.Arcs — Type

Arcs type with points [deg]. A minimum of two points are necessary for a string of line sections

Missing docstring.

Missing docstring for MultiArcs. Check Documenter's build log for details.

SphericalGeometry.Polygon — Type

Polygon type with points [deg]. A minimum of three different points are necessary for a polygon.

Missing docstring.

Missing docstring for MultiPolygons. Check Documenter's build log for details.

Functions

SphericalGeometry.along_line_angular_distance — Method

along_line_angular_distance(angular_distance₁₃::Float64,
angular_distance_line3::Float64)

The along_line_angular_distance from the start point₁ [deg] to the closest point on the great circle line to point₃. As input we need the angular_distance₁₃ [deg] between point₁ and point₃ [deg] and the closest distance angular_distance_line3 [deg] between the great circle line and point3 [deg]. The distance is non-directional.

Source: www.movable-type.co.uk/scripts/latlong.html

SphericalGeometry.along_line_angular_distance — Method

along_line_angular_distance(point₃::Point, line₁::Line)

The along_line_angular_distance from the start point₁ [deg] to the closest point on the great circle line line₁ to point₃. The distance is non-directional.

Source: www.movable-type.co.uk/scripts/latlong.html

SphericalGeometry.angular_distance — Function

angular_distance(point₃::Point, point₁::Point, point₂::Point)

Return the angular_distance [deg] from point₃ [deg] to the closest point on a great circle line section starting in point₁ [deg] and ending in point₂ [deg]. The great circle line section does not continue around the unit sphere.

The angular_distance does not change sign when being left or right of the arc.

SphericalGeometry.angular_distance — Method

angular_distance(angular_distance₁₃::Float64, azimuth₁₃::Float64,
azimuth₁₂::Float64)

Return the angular_distance [deg] from point₃ [deg] to the closest point on a great circle line starting in point₁ [deg] with azimuth₁₂ [deg]. For the calculation we need the angular_distance₁₃ and azimuth₁₃ between point₁ and point₃ [deg], and the azimuth₁₂ from point₁ to point₂ along the great circle line. It is assumed that the great circle line does not stop in point₂, but continuous around the unit sphere.

Source: www.movable-type.co.uk/scripts/latlong.html

SphericalGeometry.angular_distance — Method

angular_distance(point₃::Point, arcs::Arcs)

Return the angular_distance [deg] from point₃ [deg] to the closest point on a an arcs.

The angular_distance does not change sign when being left or right of the arc.

SphericalGeometry.angular_distance — Method

angular_distance(point₃::Point, arc₁₂::Arc)

Return the angular_distance [deg] from point₃ [deg] to the closest point on a great circle line section starting in point₁ [deg] and ending in point₂ [deg]. The great circle line section does not continue around the unit sphere.

SphericalGeometry.angular_distance — Method

angular_distance(point₃::Point, line₁::Line)

Return the angular_distance [deg] from point₃ [deg] to the closest point on a great circle line starting in point₁ [deg] with azimuth₁₂ [deg]. A positive value indicates being right of the line, and negative being left of the line.

Source: www.movable-type.co.uk/scripts/latlong.html

SphericalGeometry.angular_distance — Method

angular_distance(point₃::Point, point₁::Point, azimuth₁₂::Float64)

Return the angular_distance [deg] from point₃ [deg] to the closest point on a great circle line starting in point₁ [deg] with azimuth₁₂ [deg] from point₁ to point₂. It is assumed that the great circle line does not stop in point₂, but continuous around the unit sphere. A positive value indicates being right of the line, and negative being left of the line.

Source: www.movable-type.co.uk/scripts/latlong.html

SphericalGeometry.angular_distance — Method

angular_distance(point₁::Point, point₂::Point)

Return the angular_distance [deg] of the great circle line between the positions point₁ [deg] and point₂ [deg] on a unit sphere.

Source: www.movable-type.co.uk/scripts/latlong.html

SphericalGeometry.angular_distance — Method

angular_distance(point₃::Point, polygon::Polygon)

Return the angular_distance [deg] from point₃ [deg] to the closest point on the border of a polygon.

The angular_distance does not change sign when being left or right of the arc.

SphericalGeometry.angular_distance — Method

angular_distance(point₃::Point, points::Vector{Points{T}}) where T<:Float64

Return the angular_distance [deg] from point₃ [deg] to the closest point on a an set of points representing an set of arcs.

The angular_distance does not change sign when being left or right of the arc.

SphericalGeometry.angular_length — Method

angular_length(arcs::Arcs)

Return the angular_length [deg] along the arcs on a unit sphere.

SphericalGeometry.angular_length — Method

angular_length(line::Line)

Return the angular_length [deg] of the great circle line section on a unit sphere.

SphericalGeometry.angular_length — Method

angular_length(points::Vector{Points{Float64}})

Return the angular_length [deg] along the subsequent points on a unit sphere.

SphericalGeometry.angular_length — Method

angular_length(polygon::Polygon)

Return the angular_length deg along the polygon on a unit sphere.

SphericalGeometry.azimuth — Method

azimuth(line::Line)

Return the initial azimuth [deg], measured clockwise from the north direction, of the line on a unit sphere.

SphericalGeometry.azimuth — Method

azimuth(point₁::Point, point₂::Point)

Return the initial azimuth [deg], measured clockwise from the north direction, of the great circle line between the positions point₁ and point₂ [deg] on a unit sphere.

Source: www.movable-type.co.uk/scripts/latlong.html

SphericalGeometry.destination_point — Method

destination_point(line::Line, angular_distance::Float64)

Given a line with a startpoint and azimuth, calculate the `destina­tionpoint[deg] at theangular_distance`.

SphericalGeometry.destination_point — Method

destination_point(start_point::Point, angular_distance::Float64,
azimuth::Float64)

Given a start_point [deg], initial azimuth (clockwise from North) [deg], angular_distance [deg] calculate the position of the destina­tion_point [deg] traversing along a great circle line.

Source: www.movable-type.co.uk/scripts/latlong.html

SphericalGeometry.final_azimuth — Method

final_azimuth(line::Line)

Return the final_azimuth [deg] of the line on a unit sphere.

SphericalGeometry.final_azimuth — Method

final_azimuth(point₁::Point, point₂::Point)

Return the final azimuth [deg], measured clockwise from the north direction, of the great circle line between the positions point₁ and point₂ [deg] on a unit sphere.

Source: www.movable-type.co.uk/scripts/latlong.html

SphericalGeometry.intermediate_point — Method

intermediate_point(arc::Arc, fraction::Float64)

Return the intermediate_point [deg] at any fraction along the great circle arc [deg]. The fraction along the great circle arc is such that fraction = 0.0 is at the start of the arc and fraction 1.0 is at the end of the arc.

SphericalGeometry.intermediate_point — Method

intermediate_point(arcs::Arcs, fraction::Float64)

Return the intermediate_point [deg] at any fraction along the great circle arcs [deg]. The fraction along the great circle arcs is such that fraction = 0.0 is at the start of the arcs and fraction 1.0 is at the end of the arcs.

SphericalGeometry.intermediate_point — Method

intermediate_point(point₁::Point, point₂::Point, fraction::Float64)

Return the intermediate_point [deg] at any fraction along the great circle line section between the two points point₁ and point₂ [deg]. The fraction along the great circle line section is such that fraction = 0.0 is at point₁ [deg] and fraction 1.0 is at point₂ [deg].

Source: www.movable-type.co.uk/scripts/latlong.html

SphericalGeometry.intersection_point — Method

intersection_point(point₁::Point, point₂::Point, point₃::Point, point₄::Point)

Return the intersection point [deg] of two great circle line section given by two sets of two points: point₁ and point₂ [deg] and point₃ and point₄ [deg].

Under certain circumstances the results can be an ∞ or ambiguous solution.

SphericalGeometry.intersection_point — Method

intersection_point(arc₁::Arc, arc₂::Arc)

Return the intersection point [deg] of two great circle line sections.

Under certain circumstances the results can be an ∞ or ambiguous solution.

SphericalGeometry.intersection_point — Method

intersection_point(line₁::Line, line₂::Line)

Return the intersection point point₃ [deg] of two great circle lines 'line₁' and line₂.

Under certain circumstances the results can be an ∞ or ambiguous solution.

SphericalGeometry.intersection_point — Method

intersection_point(point₁::Point, point₂::Point, azimuth₁₃::Float64,
azimuth₂₃::Float64)

Return the intersection point point₃ [deg] of two great circle lines given by two start points [deg] point₁ and point₂ [deg] and two azimuths [deg] from point₁ to point₃ [deg], and from point₂ to point₃ [deg].

Under certain circumstances the results can be an ∞ or ambiguous solution.

Source: edwilliams.org/avform.htm

SphericalGeometry.intersection_points — Method

intersection_points(arcs₁::Arcs, arcs₂::Arcs)

Return the intersection points [deg] of two line sections.

SphericalGeometry.intersection_points — Method

intersection_points(arcs::Arcs, polygon::Polygon)

Return the intersection points [deg] of a arcs with a polygon.

SphericalGeometry.intersection_points — Method

intersection_points(polygon₁::Polygon, polygon₂::Polygon)

Return the intersection points [deg] of two polygons.

SphericalGeometry.isinside — Method

isinside(arc::Arc, polygon::Polygon)

Determine if the arc is fully inside the polygon. The border is outside the polygon.

SphericalGeometry.isinside — Method

isinside(arcs::Arcs, polygon::Polygon)

Determine if the arcs is fully inside the polygon. The border is outside the polygon.

SphericalGeometry.isinside — Method

isinside(point::Point, polygon::Polygon)

Determine if the point is inside the polygon. The border is outside the polygon.

Source: M. Bevis and J.L. Chatelain, "Locating a point on a spherical surface relative to a spherical polygon" 1989

SphericalGeometry.isinside — Method

isinside(polygon₁::Polygon, polygon₂::Polygon)

Determine if the polygon₁ is fully inside the polygon₂. The border is outside the polygon₁.

SphericalGeometry.ison — Function

ison(point::Point, arc::Arc [,tolerance::Float64=tolerance_deg])

Determine if the point is on the arc.

SphericalGeometry.ison — Function

ison(point::Point, arc::Arcs [,tolerance::Float64=tolerance_deg])

Determine if the point is on the arcs.

SphericalGeometry.ison — Function

ison(point::Point, line::Line [,tolerance::Float64=tolerance_deg])

Determine if the point is on the line.

SphericalGeometry.isonborder — Function

isonborder(point::Point, polygon::Polygon [,tolerance::Float64=tolerance_deg])

Determine if the point is on the border of the polygon.

SphericalGeometry.isselfintersecting — Method

isselfintersecting(arcs::Arcs)

SphericalGeometry.isselfintersecting — Method

isselfintersecting(points::Vector{Point{Float64}})

SphericalGeometry.isselfintersecting — Method

isselfintersecting(polygon::Polygon)

SphericalGeometry.midpoint — Method

midpoint(arcs::Arcs)

Return the half-way point midpoint [deg] on the great circle arcs [deg] on a unit sphere.

SphericalGeometry.midpoint — Method

midpoint(arc::Arc)

Return the half-way point midpoint [deg] on the great circle arc [deg] on a unit sphere.

SphericalGeometry.midpoint — Method

midpoint(point₁::Point, point₂::Point)

Return the half-way point midpoint [deg] on the great circle line between the positions point₁ and point₂ [deg] on a unit sphere.

Source: www.movable-type.co.uk/scripts/latlong.html